moved CasServerError to services
