+++ src/utils.ts
@@ -2,7 +2,7 @@ import { Notification } from '@jupyterlab/apputils';
 
 import { TreeResponse, TreeChildResponse } from './schema/types';
 import { ITreeData, ITreeChildData, TreeChildren } from './core';
-import { URLExt } from '@jupyterlab/coreutils';
+
 
 export function treeChildrenToData(children: {
   [id: string]: TreeChildResponse;
@@ -67,26 +67,4 @@ export function warnError(notifyMessage: string, logMessage?: string): void {
   }
 
   console.warn('Cassini - ' + logMessage);
-}
-
-
-export class CasServerError extends Error {
-  endpoint: string;
-  query: string | null;
-  info: string | null;
-
-  constructor(reason: string, url: string, info?: string ) {
-    super(reason);
-    const { pathname, search } = URLExt.parse(url);
-    
-    this.endpoint = pathname;
-    this.query = search || null;
-    this.info = info || null;
-  }
-
-  notify(): void {
-    const notifyMessage = `${this.endpoint}${this.query}, returned ${this.message}, check out browser and server log for more details.`;
-    const logMessage = `Cassini server error ${this.message} at ${this.endpoint}${this.query}, caused by: \n\n ${this.info}`;
-    warnError(notifyMessage, logMessage);
-  }
 }
